Agent: fix has_role_notifications property (#773)
